all obvious bugs fixed.
