+++ quspin/operators/__init__.py
@@ -1,3 +1,6 @@
-from .ops import *
+from .hamiltonian import *
+from .ops_dict import *
+from .exp_op import *
+from .HamiltonianOperator import *
